Fixed: CustomFormat size specs in already grabbed check
Sizes need to be parsed as a long not an int else anything with a size
> 2GiB will fail to be parsed and be set with size 0

Fixes #4262

+++ src/NzbDrone.Core/Parser/Augmenters/AugmentWithHistory.cs
@@ -44,7 +44,7 @@ namespace NzbDrone.Core.Parser.Augmenters
                     releaseInfo.IndexerId = indexerId;
                 }
 
-                if (int.TryParse(history.Data.GetValueOrDefault("size"), out var size))
+                if (long.TryParse(history.Data.GetValueOrDefault("size"), out var size))
                 {
                     releaseInfo.Size = size;
                 }
